fix new/type page crash due to ssr function export (close #63)
